{{cookiecutter.project_slug}}/requirements/base.txt
argon2-cffi==21.3.0
pytz==2022.6
python-dotenv==1.0.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
nanoid==2.0.0
whitenoise==6.2.0  # https://github.com/evansd/whitenoise

# Django
# ------------------------------------------------------------------------------
Django==4.1.3
django-cors-headers==3.13.0
django-environ==0.9.0
django-charid-field==0.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt[crypto]==5.2.2
djangorestframework-camel-case==1.3.0
